Adding input_boy.py, updating grailed_bot.py, adding newest_tweet_id.txt to store latest tweet id
input_bot.py will serve the same features as grailed_bot.py, however it will allow for users to input a twitter handle and a keyword to get recent tweets from. grailed_bot.py will take up to 4 parameters. current status of scripts: modifying code to prevent continuous updating of newest_tweet_id.txt when script is run, as this overrwrites previous output and causes the output to have no tweets (since the .txt file will contain the most recent tweet, which is treated as the start tweet. script should have the start tweet as the most recent tweet from the previous instance of the script running, not the current instance